remove obsolete files and rename utils
