--- python/start.py
@@ -1,5 +1,0 @@
-import sys
-# Currently hard-wire location of generated files. This will need to change!
-sys.path.append('../PETSIRD/python/')
-import prd
-
